planner.cleaning_buffer_counter
we now check the planner to see if its currently clearing the buffer from a M410 (quick stop) call, if so we quit the function.

this allows for the FREEZE switch to stop the machine smoothly followed by a M410 to clear the buffer allowing us to continue machining without having to dial in our part again.